requirements.txt
requests
beautifulsoup4
transformers
torch
scikit-learn
spacy
PyPDF2
python-docx
